Refs #36065 -- Extracted composite primary key order by tests.
