+++ tests/composite_pk/test_filter.py
@@ -70,30 +70,6 @@ class CompositePKFilterTests(TestCase):
             ):
                 Comment.objects.filter(text__gt=expr).count()
 
-    def test_order_comments_by_pk_asc(self):
-        self.assertSequenceEqual(
-            Comment.objects.order_by("pk"),
-            (
-                self.comment_1,  # (1, 1)
-                self.comment_2,  # (1, 2)
-                self.comment_3,  # (1, 3)
-                self.comment_5,  # (1, 5)
-                self.comment_4,  # (2, 4)
-            ),
-        )
-
-    def test_order_comments_by_pk_desc(self):
-        self.assertSequenceEqual(
-            Comment.objects.order_by("-pk"),
-            (
-                self.comment_4,  # (2, 4)
-                self.comment_5,  # (1, 5)
-                self.comment_3,  # (1, 3)
-                self.comment_2,  # (1, 2)
-                self.comment_1,  # (1, 1)
-            ),
-        )
-
     def test_filter_comments_by_pk_gt(self):
         c11, c12, c13, c24, c15 = (
             self.comment_1,
